Encode and decode execution provider in the UI
